refactor: reorganize react-pi src into runtime/, client/, and node/ layers
Replace the flat pi-prefixed file layout with folders that encode the
package's layering: runtime/ (React-facing, browser-safe), client/
(PiClient transports, browser-safe), node/ (the only folder allowed to
import @earendil-works/pi-* at runtime), with the shared JSON-safe
mirror at src/types.ts. Folder names provide the namespace, so the pi
file-name prefixes are dropped; public exports and entry points are
unchanged.
